story-universe/chronicle-keeper/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
aiosqlite>=0.19.0
pyzmq>=25.1.0
pyyaml>=6.0.1
